refactor: validation 형식 수정
- validate 함수 제네릭 타입으로 변경
